pyproject.toml
[tool.poetry]
name = "simple-ssh-client"
version = "1.0.0"
description = "Simple SSH client based on Paramiko"
authors = ["Wojciech Richert <33549959+wojtekrichert@users.noreply.github.com>"]
keywords = ["ssh", "paramiko"]
readme = "README.md"
homepage = "https://github.com/wojtekrichert/simple-ssh-client"
repository = "https://github.com/wojtekrichert/simple-ssh-client"

[tool.poetry.dependencies]
python = "^3.9"
paramiko = "^2.9.2"
retry = "^0.9.2"
pydantic = "^1.9.0"

[tool.poetry.dev-dependencies]
factory-boy = "^3.2.1"
pytest = "^7.0.1"
# code quality
black = "^21.11b1"
isort = "==5.10.1"
flake8 = "==4.0.1"
flake8-isort = "==4.1.1"
pre-commit = "^2.16.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
